+++ asn1_group/B168_cb_dcall.h
@@ -1,4 +1,4 @@
 #include <stdint.h>
 #include <stdio.h>
 
-void decode_B168(uint8_t *hex_data, size_t length, size_t *index);
+void decode_B168(uint8_t *hex_data, size_t length, int *index);
